Use 16-bit size for ResourceKeys
Internally, ResourceKey required the size to fit into a uint16_t so this
makes that explicit in the public API. It also changes how the size is
stored to instead record the num32DataCount directly and then convert to
bytes as needed, whereas previously it was requiring that the actual
byte count fit into a uint16_t. This gives a bit more head room.

Call sites to the ResourceKey builders are updated to now have the
responsibility of checking that their size can fit into a uint16_t. For
the most part, these were fixed or trivially small variable key sizes.
The two exceptions were Ganesh's style key (with dashes) and its
inherited key system for shapes with applied styles and path effects.
They now have reasonable limits to prevent the keys from growing bigger
than about 1kb.

Bug: b/495700484
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1199497
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
(cherry picked from commit 0566b2f5f0d1f218d9990eba838af16826d2e7e3)
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1206376
Commit-Queue: Thomas Smith <thomsmit@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>

+++ src/gpu/graphite/GraphiteResourceKey.h
@@ -46,7 +46,7 @@ public:
 
     class Builder : public ResourceKey::Builder {
     public:
-        Builder(GraphiteResourceKey* key, ResourceType type, int data32Count)
+        Builder(GraphiteResourceKey* key, ResourceType type, uint16_t data32Count)
                 : ResourceKey::Builder(key, type, data32Count) {}
     };
 };
